backend/requirements.txt
fastapi[all]
uvicorn[standard]==0.29.0
huggingface_hub==0.23.0
python-multipart==0.0.9  # necesario si luego permites subida de archivos
python-dotenv
pydantic
pydantic-settings
nbformat
numpy
imageio
scipy
matplotlib
ipython
aiofiles
python-multipart
google-api-python-client
google-auth
authlib